Form Logic Issue for Product Whitelists.
* A little patch to solve an issue that caused the whitelist properties
of a `Product` object not to be updated in a very singular case. If
there was at least a single element in the `product_whitelist` or
`platform_whitelist` properties, if in an update the HTML form
check-box elements for these properties were all unchecked, the `params`
with the POST request would send empty arrays, causing the issue.
Solved by simply sending `0` (zeroes) instead of `nil` objects for
unchecked elements, and then wiping these zeroes from the array on a
`#before_validation` callback on the model.

* There's a little fix for the Capistrano configuration as well.