Spinner Stylesheet Duplication Fix.
This commit fixes #12, which addresses an issue related to multiple
copies of the same `spinner.css.scss` file across the stylesheet files'
tree.
The fix is a simple modification of the `application.html.haml` file so
that it always loads the contents of `./app/assets/stylesheets/all/` in
addition to the files particular to each terminal: `desktop`, `tablet`
and `mobile`.